requirements.txt
alembic==1.13.0
annotated-types==0.7.0
anyio==3.6.2
asttokens==2.4.1
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
Automat==0.8.0
Babel==2.13.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.1.1
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
cached-property==1.5.1
cachetools==5.3.3
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.7
cloud-init==23.3.3
colorama==0.4.6
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cssselect==1.2.0
dbus-python==1.2.16
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
distro-info==0.23+ubuntu1.1
dnspython==2.4.2
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.1
exchangelib==5.1.0
executing==2.0.1
fake-useragent==1.1.3
flask==3.0.1
Flask-Babel==2.0.0
Flask-Cors==3.0.10
Flask-Login==0.6.3
Flask-Migrate==4.0.5
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
google-ai-generativelanguage==0.1.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-aiplatform==1.44.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-secret-manager==2.19.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.1.0rc1
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==0.17.2
httplib2==0.14.0
httpx==0.24.1
hyperlink==19.0.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
incremental==16.10.1
ipython==8.12.3
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.6.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.9.2
Mako==1.3.0
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.24.2
oauthlib==3.1.0
openai==1.51.2
openpyxl==3.0.10
outcome==1.2.0
packaging==24.0
pandas==1.5.3
parameterized==0.9.0
parse==1.19.0
parso==0.8.3
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.11.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.11.1
pydantic==2.9.2
pydantic-core==2.23.4
pyee==8.2.2
pygments==2.17.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.5.0
PyNaCl==1.3.0
pyngrok==7.0.1
pyOpenSSL==19.0.0
pyppeteer==1.0.2
pyquery==2.0.0
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
pyspnego==0.10.2
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-docx==0.8.11
python-dotenv==1.0.1
python-pptx==0.6.21
python-telegram-bot==20.3
pytz==2022.7.1
PyYAML==5.3.1
redis==5.0.1
requests==2.29.0
requests-html==0.10.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
SecretStorage==2.3.1
selenium==4.9.1
service-identity==18.1.0
shapely==2.0.3
simplejson==3.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sos==4.5.6
soupsieve==2.4.1
SQLAlchemy==2.0.25
ssh-import-id==5.10
stack-data==0.6.3
systemd-python==234
telegram-send==0.34
texttable==1.6.2
tomli==2.0.1
tqdm==4.65.0
traitlets==5.14.1
trio==0.22.0
trio-websocket==0.10.2
Twisted==18.9.0
typing-extensions==4.12.2
tzdata==2023.4
tzlocal==5.2
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.15
w3lib==2.1.1
wadllib==1.3.3
wcwidth==0.2.13
websocket-client==0.53.0
websockets==10.4
werkzeug==3.0.1
wsproto==1.2.0
wtforms==3.1.2
XlsxWriter==3.0.9
yapf==0.40.2
zipp==3.15.0
zope.interface==4.7.1
